requirements.txt
fastapi==0.115.8
jwcrypto==1.5.6
pyjwt==2.10.1
